Bugfix display for MEF with only 1 extension; Bugfix overlap montage.mosaic call; Add oversampling option to solveAstrometry;

+++ misc/display.py
@@ -4,7 +4,7 @@
 #
 # display.py
 #
-# Last update 09/10/2019
+# Last update 21/01/2020
 #
 ################################################################################
 """
@@ -15,6 +15,9 @@
 import os
 from distutils import spawn
 import time
+import argparse
+import sys
+
 
 # papi
 from misc.paLog import log
@@ -39,6 +42,7 @@ frame_no = 0
 created_frames = 0
 MAX_FRAMES_NO = 10
 
+
 ################################################################################
 # Launch the DS9 display
 def startDisplay():
@@ -59,79 +63,77 @@ def startDisplay():
         if stdout_handle.read() == 'no\n':
             time.sleep(3)
         time.sleep(1)
-        #for i in range(0, MAX_FRAMES_NO-1): # when ds9 start, it has already one frame
-        #    os.system(("%s/xpaset -p ds9 frame new" % xpa_path))
-            
+
     else:
-        pass
         # DS9 is already running...
-        #print "Warning, display is already running"
-        #os.system(("%s/xpaset -p ds9 frame delete all" % xpa_path))
-      
-    
+        pass
+
+
 ################################################################################
 # Show the current frame into DS9 display
-
+# ##############################################################################
 def showFrame(frame, del_all=False):
-  """
-  Show in DS9 display the file/s given in the input
-  """
+    """
+    Show in DS9 display the file/s given in the input
+    """
   
-  # Check display
-  startDisplay()
+    # Check display
+    startDisplay()
   
-  global frame_no
-  global created_frames
+    global frame_no
+    global created_frames
   
-  # frame could be a single file or a file list 
-  if type(frame)==type(list()): 
+    # frame could be a single file or a file list
+    if type(frame) == type(list()):
         fileList = frame
         os.system(("%s/xpaset -p ds9 frame delete all" % (xpa_path)))
         delete_all = False
-  elif os.path.isfile(frame):
+    elif os.path.isfile(frame):
         delete_all = del_all
         fileList = [frame]
-  
-  for file in fileList:
+
+    for file in fileList:
         f = datahandler.ClFits(file)
-        if f.mef:
+        if f.mef and f.getNExt() > 1:
                 # Multi-Extension FITS files
                 if f.getInstrument() == 'hawki':
-                    # HAWK-I Dark files don't have WCS information required by ds9 mosaicimage
-                    # PANIC  Dark files do have WCS information.  
+                    # HAWK-I Dark files don't have WCS information required by
+                    # ds9 mosaicimage.
+                    # PANIC  Dark files do have WCS information.
                     if delete_all: os.system(("%s/xpaset -p ds9 frame delete all" % (xpa_path)))
                     os.system(("%s/xpaset -p ds9 frame new" % xpa_path))
                     os.system(("%s/xpaset -p ds9 cmap Heat" % xpa_path))
                     os.system(("%s/xpaset -p ds9 scale zscale" % xpa_path ))
                     os.system(("%s/xpaset -p ds9 file multiframe %s" % (xpa_path, file)))
-                    #os.system(("%s/xpaset -p ds9 medatacube multiframe %s" % (xpa_path, file)))
+                    # os.system(("%s/xpaset -p ds9 medatacube multiframe %s" % (xpa_path, file)))
                 else:
                     # Beware, 'mosaicimage' ds9 facility require WCS information
-                    if delete_all: os.system(("%s/xpaset -p ds9 frame delete all" % (xpa_path)))
-                    if frame_no<MAX_FRAMES_NO:
-                        if created_frames<(frame_no+1):
+                    if delete_all:
+                        os.system(("%s/xpaset -p ds9 frame delete all" % (xpa_path)))
+                    if frame_no < MAX_FRAMES_NO:
+                        if created_frames < (frame_no+1):
                             os.system(("%s/xpaset -p ds9 frame new" % xpa_path))
-                            created_frames+=1
+                            created_frames += 1
                         os.system(("%s/xpaset -p ds9 frame frameno %d" % (xpa_path, frame_no+1)))
-                        frame_no+=1
+                        frame_no += 1
                     else:
                         frame_no = 1
                         os.system(("%s/xpaset -p ds9 frame frameno %d" % (xpa_path, frame_no)))
-                    os.system(("%s/xpaset -p ds9 single" % xpa_path)) 
+                    os.system(("%s/xpaset -p ds9 single" % xpa_path))
                     os.system(("%s/xpaset -p ds9 file mosaicimage %s" % (xpa_path, file)))
                     os.system(("%s/xpaset -p ds9 cmap Heat" % xpa_path))
                     os.system(("%s/xpaset -p ds9 scale zscale" % xpa_path ))
                     os.system(("%s/xpaset -p ds9 zoom to fit" % xpa_path))
         else:
                 # Single FITS files
-                if delete_all: os.system(("%s/xpaset -p ds9 frame delete all" % (xpa_path)))
-                if frame_no<MAX_FRAMES_NO:
-                    #os.system(("%s/xpaset -p ds9 frame new" % xpa_path))
+                if delete_all:
+                    os.system(("%s/xpaset -p ds9 frame delete all" % (xpa_path)))
+                if frame_no < MAX_FRAMES_NO:
                     os.system(("%s/xpaset -p ds9 frame frameno %d" % (xpa_path, frame_no+1)))
-                    frame_no+=1
+                    frame_no += 1
                 else:
-                    frame_no=1
-                    os.system(("%s/xpaset -p ds9 frame frameno %d" % (xpa_path, frame_no))) 
+                    frame_no = 1
+                    os.system(("%s/xpaset -p ds9 frame frameno %d" % (xpa_path, frame_no)))
                 os.system(("%s/xpaset -p ds9 single" % xpa_path))
                 os.system(("%s/xpaset -p ds9 cmap Heat" % xpa_path))
                 os.system(("%s/xpaset -p ds9 file %s" %(xpa_path, file)))
@@ -143,28 +145,27 @@ def showFrame(frame, del_all=False):
 # Show the current frame into DS9 display
 
 def showSingleFrames(framelist):
-  """ Display a single frame, not supposing a MEF file """
-  
-  global next_frameno
+    """ Display a single frame, not supposing a MEF file """
 
-  #Check display
-  startDisplay()
-  
-  nframes=len(framelist)
-  for i in range(nframes):
-      #os.system(("%s/xpaset -p ds9 frame new" % xpa_path))
-      os.system(("%s/xpaset -p ds9 frame frameno %d" % (xpa_path, next_frameno)))
-      os.system(("%s/xpaset -p ds9 frame reset" % xpa_path))
-      os.system(("%s/xpaset -p ds9 tile" % xpa_path))
-      os.system(("%s/xpaset -p ds9 cmap Heat" % xpa_path))
-      os.system(("%s/xpaset -p ds9 file %s" %(xpa_path, framelist[i])))
-      os.system(("%s/xpaset -p ds9 scale zscale" % xpa_path ))
-      os.system(("%s/xpaset -p ds9 zoom to fit" % xpa_path))
-      
-      if next_frameno == 4:
-          next_frameno=1
-      else:
-          next_frameno=next_frameno+1
+    global next_frameno
+
+    # Check display
+    startDisplay()
+
+    nframes = len(framelist)
+    for i in range(nframes):
+        os.system(("%s/xpaset -p ds9 frame frameno %d" % (xpa_path, next_frameno)))
+        os.system(("%s/xpaset -p ds9 frame reset" % xpa_path))
+        os.system(("%s/xpaset -p ds9 tile" % xpa_path))
+        os.system(("%s/xpaset -p ds9 cmap Heat" % xpa_path))
+        os.system(("%s/xpaset -p ds9 file %s" %(xpa_path, framelist[i])))
+        os.system(("%s/xpaset -p ds9 scale zscale" % xpa_path ))
+        os.system(("%s/xpaset -p ds9 zoom to fit" % xpa_path))
+
+        if next_frameno == 4:
+            next_frameno = 1
+        else:
+            next_frameno = next_frameno + 1
           
     
 ################################################################################
@@ -195,10 +196,20 @@ def showPDF(filename):
     else:
         log.info("Give file %s does not look a PDF file" % filename)
 
+
 ################################################################################
 if __name__ == "__main__":
+
+    parser = argparse.ArgumentParser()
+    parser.add_argument('-f', '--file', dest='file',
+                        help='name of file to display')
+
+    args = parser.parse_args()
+    if not args.file:
+        parser.print_usage()
+        sys.exit(0)
+
     # test
     startDisplay()
-    f = "../data_test/panic.20191008_0001.fits"
-    showFrame(f)
+    showFrame(args.file)
